+++ spec/unit/recipes/default_spec.rb
@@ -3,13 +3,22 @@ require 'spec_helper'
 describe 'memcached::default' do
   let(:chef_run) { ChefSpec::Runner.new.converge(described_recipe) }
 
-  it 'installs the libmemcache-dev package' do
-    expect(chef_run).to install_package('libmemcache-dev')
+  it 'installs the memcached package' do
+    expect(chef_run).to install_package('memcached')
+  end
+
+  it 'creates memcached service with :nothing action' do
+    expect(chef_run).to_not start_service('memcached')
+    expect(chef_run).to_not enable_service('memcached')
   end
 
   context 'on rhel' do
-    let(:chef_run) { ChefSpec::Runner.new(:platform => 'redhat', :version => '6.3').converge(described_recipe) }
+    let(:chef_run) { ChefSpec::Runner.new(platform: 'redhat', version: '6.4').converge(described_recipe) }
     let(:template) { chef_run.template('/etc/sysconfig/memcached') }
+
+    it 'installs libmemcached-devel package'do
+      expect(chef_run).to install_package('libmemcached-devel')
+    end
 
     it 'writes the /etc/sysconfig/memcached' do
       expect(template).to be
@@ -23,17 +32,13 @@ describe 'memcached::default' do
     end
   end
 
-  context 'on smartos' do
-    let(:chef_run) { ChefSpec::Runner.new(:platform => 'smartos', :version => 'joyent_20130111T180733Z').converge(described_recipe) }
-
-    it 'enables the memcached service' do
-      expect(chef_run).to enable_service('memcached')
-    end
-  end
-
   context 'on ubuntu' do
-    let(:chef_run) { ChefSpec::Runner.new(:platform => 'ubuntu', :version => '12.04').converge(described_recipe) }
+    let(:chef_run) { ChefSpec::Runner.new(platform: 'ubuntu', version: '12.04').converge(described_recipe) }
     let(:template) { chef_run.template('/etc/memcached.conf') }
+
+    it 'installs libmemcache-dev package'do
+      expect(chef_run).to install_package('libmemcache-dev')
+    end
 
     it 'writes the /etc/memcached.conf' do
       expect(template).to be
@@ -46,4 +51,5 @@ describe 'memcached::default' do
       expect(template).to notify('service[memcached]').to(:restart)
     end
   end
+
 end